historical-analysis/requirements.txt
appdirs==1.4.4
py-radix==0.10.0
mrtparse==2.1.0
portion==2.2.0
pybgpstream==2.0.2
requests==2.27.1
pandas==1.1.5
